Move platform_version and platform_shortname calculations into Package
