+++ spec/unit/package_spec.rb
@@ -30,6 +30,99 @@ module Omnibus
 
         described_class.generate(package, data)
       end
+
+      describe '#arch' do
+        it 'returns the architecture' do
+          stub_ohai(platform: 'ubuntu', version: '12.04') do |data|
+            data['kernel']['machine'] = 'x86_64'
+          end
+          expect(described_class.arch).to eq('x86_64')
+        end
+      end
+
+      describe '#platform_shortname' do
+        it 'returns el on rhel' do
+          stub_ohai(platform: 'redhat', version: '6.4')
+          expect(described_class.platform_shortname).to eq('el')
+        end
+
+        it 'returns #platform on non-rhel systems' do
+          stub_ohai(platform: 'ubuntu', version: '12.04')
+          expect(described_class.platform_shortname).to eq('ubuntu')
+        end
+      end
+
+      describe '#platform_version' do
+        shared_examples 'a version manipulator' do |platform_shortname, version, expected|
+          context "on #{platform_shortname}-#{version}" do
+            it 'returns the correct value' do
+              stub_ohai(platform: 'ubuntu', version: '12.04') do |data|
+                data['platform'] = platform_shortname
+                data['platform_version'] = version
+              end
+
+              expect(described_class.platform_version).to eq(expected)
+            end
+          end
+        end
+
+        it_behaves_like 'a version manipulator', 'arch', '2009.02', '2009.02'
+        it_behaves_like 'a version manipulator', 'arch', '2014.06.01', '2014.06'
+        it_behaves_like 'a version manipulator', 'debian', '7.1', '7'
+        it_behaves_like 'a version manipulator', 'debian', '6.9', '6'
+        it_behaves_like 'a version manipulator', 'ubuntu', '10.04', '10.04'
+        it_behaves_like 'a version manipulator', 'ubuntu', '10.04.04', '10.04'
+        it_behaves_like 'a version manipulator', 'fedora', '11.5', '11'
+        it_behaves_like 'a version manipulator', 'freebsd', '10.0', '10'
+        it_behaves_like 'a version manipulator', 'rhel', '6.5', '6'
+        it_behaves_like 'a version manipulator', 'el', '6.5', '6'
+        it_behaves_like 'a version manipulator', 'centos', '5.9.6', '5'
+        it_behaves_like 'a version manipulator', 'aix', '7.1', '7.1'
+        it_behaves_like 'a version manipulator', 'gentoo', '2004.3', '2004.3'
+        it_behaves_like 'a version manipulator', 'mac_os_x', '10.9.1', '10.9'
+        it_behaves_like 'a version manipulator', 'openbsd', '5.4.4', '5.4'
+        it_behaves_like 'a version manipulator', 'slackware', '12.0.1', '12.0'
+        it_behaves_like 'a version manipulator', 'solaris2', '5.9', '5.9'
+        it_behaves_like 'a version manipulator', 'suse', '5.9', '5.9'
+        it_behaves_like 'a version manipulator', 'omnios', 'r151010', 'r151010'
+        it_behaves_like 'a version manipulator', 'smartos', '20120809T221258Z', '20120809T221258Z'
+        it_behaves_like 'a version manipulator', 'windows', '5.0.2195', '2000'
+        it_behaves_like 'a version manipulator', 'windows', '5.1.2600', 'xp'
+        it_behaves_like 'a version manipulator', 'windows', '5.2.3790', '2003r2'
+        it_behaves_like 'a version manipulator', 'windows', '6.0.6001', '2008'
+        it_behaves_like 'a version manipulator', 'windows', '6.1.7600', '7'
+        it_behaves_like 'a version manipulator', 'windows', '6.1.7601', '2008r2'
+        it_behaves_like 'a version manipulator', 'windows', '6.2.9200', '8'
+        it_behaves_like 'a version manipulator', 'windows', '6.3.9200', '8.1'
+
+        context 'given an unknown platform' do
+          before do
+            stub_ohai(platform: 'ubuntu', version: '12.04') do |data|
+              data['platform'] = 'bacon'
+              data['platform_version'] = '1.crispy'
+            end
+          end
+
+          it 'raises an exception' do
+            expect { described_class.platform_version }
+              .to raise_error(UnknownPlatform)
+          end
+        end
+
+        context 'given an unknown windows platform version' do
+          before do
+            stub_ohai(platform: 'ubuntu', version: '12.04') do |data|
+              data['platform'] = 'windows'
+              data['platform_version'] = '1.2.3'
+            end
+          end
+
+          it 'raises an exception' do
+            expect { described_class.platform_version }
+              .to raise_error(UnknownPlatformVersion)
+          end
+        end
+      end
     end
 
     describe '.for_package' do
@@ -51,22 +144,15 @@ module Omnibus
         expect(instance[:platform]).to eq('ubuntu')
       end
 
-      it 'ensures platform version is properly truncated' do
-        allow(File).to receive(:read).and_return('{ "platform": "el", "platform_version": "5.10" }')
-        instance = described_class.for_package(package)
-
-        expect(instance[:platform_version]).to eq('5')
-      end
-
       it 'ensures an iteration exists' do
-        allow(File).to receive(:read).and_return('{ "platform": "el", "platform_version": "5.10" }')
+        allow(File).to receive(:read).and_return('{}')
         instance = described_class.for_package(package)
 
         expect(instance[:iteration]).to eq(1)
       end
 
       it 'does not change existing iterations' do
-        allow(File).to receive(:read).and_return('{ "platform": "el", "platform_version": "5.10", "iteration": 4}')
+        allow(File).to receive(:read).and_return('{ "iteration": 4 }')
         instance = described_class.for_package(package)
 
         expect(instance[:iteration]).to eq(4)
